pyproject.toml
[tool.poetry]
name = "ethchange"
version = "0.1.0"
description = "Etherium exchange API for facilitating quick and secure transactions"
authors = ["Ommar Shaikh <ommarshaikh20@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.10, <3.12"
django = "^4.1.4"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
dynaconf = "^3.1.11"
pillow = "^9.4.0"
loguru = "^0.6.0"
web3 = "^5.31.3"
tomli = "^2.0.1"
dependency-injector = "^4.41.0"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.4"
pylint = "^2.15.10"
nox = "^2022.11.21"
pytest = "^7.2.0"
coverage = { extras = ["toml"], version = "^7.0.5" }
pytest-sugar = "^0.9.6"
pytest-cases = "^3.6.13"
pytest-django = "^4.5.2"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
runserver = "task:run_server"


[tool.isort]
py_version = 310
profile = "black"
line_length = "120"
src_paths = ["ethchange", "configs", "test"]
known_first_party = ["ethchange", "configs", "test"]
skip = [".github", "tools", "docs", "examples", ".nox", ".venv", ".pytest_cache", "dist"]
skip_gitignore = true
atomic = true
lines_after_imports = 2
ignore_comments = true
virtual_env = './venv'
remove_redundant_aliases = true
color_output = true
supported_extensions = ["py"]
add_imports = ["from __future__ import annotations", "from loguru import logger"]
group_by_package = true


[tool.black]
--check = "no"
diff = "no"
color = true
line-length = 120
exclude = '''
/(
    \.direnv
    |\.eggs
    |\.git
    |\.hg
    |\.mypy_cache
    |\.nox
    |\.tox
    |\.venv
    |venv
    |\.svn
    |\.ipynb_checkpoints
    |_build
    |buck-out
    |build
    |dist
    |__pypackages__
)/
'''
include = '\.py?$'
target-version = ["py310"]
